// backend/src/modules/rewards/schemas/reward/RewardFilters.schema.ts
import z from "zod";

export const RewardFiltersSchema = z.object({
    label: z.string().optional(),
    releaseDate: z.string().optional(),
    count: z.number().optional(),
    addition: z.string().optional(),
    description: z.string().optional(),
}).strict()
export type TypeofRewardFiltersSchema = z.infer<typeof RewardFiltersSchema>